server/models/LLM/fastapi-roleplay-project/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-core
langchain-google-genai
langchain-openai
langchain-community
faiss-cpu
pydantic